requirements.txt
asgiref==3.8.1
Django==5.1.6
sqlparse==0.5.3
tzdata==2025.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
PyJWT==2.8.0
pycryptodome==3.21.0
oscrypto==1.3.0
cryptography==43.0.1
asn1crypto==1.5.1
Pillow==10.1.0